chore(test): move oss script editor to oss folder
